+++ src/common-lib/urls/urls.test.ts
@@ -1,0 +1,19 @@
+import { isExternalHref } from "./urls";
+
+describe("urls.ts", () => {
+  describe("isExternalHref()", () => {
+    test.each([
+      "https://teachers.thenational.academy",
+      "https://classroom.thenational.academy",
+      "https://www.example.com",
+    ])("'%s' is external", (href) => {
+      expect(isExternalHref(href)).toBe(true);
+    });
+    test.each(["/about", "#email-form", "http://localhost/contact-us"])(
+      "'%s' is not external",
+      (href) => {
+        expect(isExternalHref(href)).toBe(false);
+      }
+    );
+  });
+});