+++ src/components/OakLink/OakLink.tsx
@@ -1,5 +1,6 @@
 import Link, { LinkProps } from "next/link";
-import { FC, ReactNode } from "react";
+import { forwardRef, ReactNode } from "react";
+import styled from "styled-components";
 
 import {
   isExternalHref,
@@ -7,12 +8,20 @@ import {
   resolveOakHref,
   ResolveOakHrefProps,
 } from "../../common-lib/urls";
+import flex from "../../styles/utils/flex";
 import { HTMLAnchorProps } from "../Button/common";
+import { FlexProps } from "../Flex";
 
-export type OakLinkProps = Omit<LinkProps, "href"> & {
-  children: ReactNode;
-  htmlAnchorProps?: HTMLAnchorProps;
-} & (
+const OakLinkA = styled.a`
+  ${flex}
+`;
+
+export type OakLinkProps = Omit<LinkProps, "href" | "passHref"> &
+  FlexProps & {
+    children: ReactNode;
+    className?: string;
+    htmlAnchorProps?: HTMLAnchorProps;
+  } & (
     | {
         /**
          * To encourage the ues of 'page' prop (which will get resolved to an href)
@@ -33,12 +42,27 @@ const getOakLinkHref = (props: OakLinkProps) => {
 };
 export const getOakLinkLinkProps = (props: OakLinkProps): LinkProps => {
   const href = getOakLinkHref(props);
-  const { children, htmlAnchorProps, ...linkProps } = props;
-  return { href, ...linkProps };
+
+  const { as, replace, scroll, shallow, prefetch, locale } = props;
+
+  return { href, as, replace, scroll, shallow, prefetch, locale };
 };
-export const getOakLinkAnchorProps = (props: OakLinkProps): HTMLAnchorProps => {
+export const getOakLinkAnchorProps = (
+  props: OakLinkProps
+): HTMLAnchorProps & FlexProps => {
   const href = getOakLinkHref(props);
-  const { children, htmlAnchorProps } = props;
+  const {
+    as,
+    replace,
+    scroll,
+    shallow,
+    prefetch,
+    locale,
+    children,
+    className,
+    htmlAnchorProps,
+    ...styleProps
+  } = props;
 
   const isExternal = isExternalHref(href);
   const target = isExternal ? "_blank" : undefined;
@@ -46,6 +70,8 @@ export const getOakLinkAnchorProps = (props: OakLinkProps): HTMLAnchorProps => {
   return {
     children,
     target,
+    className,
+    ...styleProps,
     ...htmlAnchorProps,
   };
 };
@@ -56,16 +82,20 @@ export const getOakLinkAnchorProps = (props: OakLinkProps): HTMLAnchorProps => {
  * It's intended to help centralise information about our url structures,
  * and to facilitate behaviour link "open in a new tab" and tracking.
  *
- * @tood add track props
+ * @todo add track props
  * @todo currently this allows href as any string, do we want to further
  * restrict it?
  */
-const OakLink: FC<OakLinkProps> = (props) => {
+const OakLink = forwardRef<HTMLAnchorElement, OakLinkProps>((props, ref) => {
   return (
-    <Link {...getOakLinkLinkProps(props)}>
-      <a {...getOakLinkAnchorProps(props)} />
+    <Link {...getOakLinkLinkProps(props)} passHref>
+      <OakLinkA ref={ref} {...getOakLinkAnchorProps(props)} />
     </Link>
   );
+});
+
+OakLink.defaultProps = {
+  $display: "inline",
 };
 
 export default OakLink;